# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RTM AI Analytics Dashboard API",
    "description": "제조업 B2B 마케팅 분석을 위한 GA4 기반 대시보드 API",
    "version": "1.0.0",
    "contact": {
      "name": "RTM AI Team",
      "email": "namgyuyooo@rtm.ai"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://ga-git-main-namgyuyooos-projects.vercel.app",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/ai-insight": {
      "get": {
        "tags": ["AI Insights"],
        "summary": "최신 AI 인사이트 조회",
        "description": "특정 타입의 최신 AI 인사이트를 조회합니다.",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "인사이트 타입",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["dashboard", "traffic", "utm-cohort", "keyword-cohort"]
            }
          },
          {
            "name": "propertyId",
            "in": "query",
            "description": "GA4 속성 ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIInsightResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["AI Insights"],
        "summary": "새 AI 인사이트 생성",
        "description": "AI를 활용하여 새로운 인사이트를 생성합니다.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAIInsightRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "인사이트 생성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIInsightResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AIInsightResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "insight": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "description": "인사이트 ID"
              },
              "type": {
                "type": "string",
                "description": "인사이트 타입"
              },
              "result": {
                "type": "string",
                "description": "AI 분석 결과"
              },
              "createdAt": {
                "type": "string",
                "format": "date-time",
                "description": "생성 일시"
              }
            }
          }
        }
      },
      "CreateAIInsightRequest": {
        "type": "object",
        "required": ["type", "propertyId"],
        "properties": {
          "type": {
            "type": "string",
            "description": "인사이트 타입",
            "enum": ["dashboard", "traffic", "utm-cohort", "keyword-cohort"]
          },
          "propertyId": {
            "type": "string",
            "description": "GA4 속성 ID"
          },
          "model": {
            "type": "string",
            "description": "사용할 AI 모델"
          },
          "template": {
            "type": "string",
            "description": "프롬프트 템플릿 ID"
          }
        }
      },
...
    }
  }
}